[ios] Base IOS app structure implementation
Creation of the main views, which are:

- The DeviceListView, where the
DeviceCardView are displayed. From it,
the user can display a QR code scanner
on a modal popup, by clicking on the +
icon, on the top right corner.

- The DeviceView, where all informations
about a specific device is meant to be
shown (currently empty).

- The ProtocolView, which display logs
about a running attestation / enrollment
(currently empty).

- The SettingsView, to control the app
behaviour (currently empty).

Creation of the DeviceCardView component,
that represent a specific Device. From it,
the user can navigate to the DeviceView
or the ProtocolView.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>