[ios] Add Device info and Device edit views
Add a Device info view, that appears when
clicking on a Device card, and that displays
informations about the device.

Add a Device edit view, that appears when clicking
the pencil button on top right of the device info
view, and which allows to change the device name, or
to delete it.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>